Add files via upload
PROBLEM STATEMENT
Analyze and visualize customer churn data for a telecom company and explore customer demographics, offers, contract types, payment methods and service usage
Identify factors that contribute to customer churn

What is the churn rate for the given dataset? Provide a pie chart or donut chart to visualize the percentage of customers who have churned versus those who have stayed.
How is the age and gender of the customers distributed in the given dataset?

What is the distribution of customer tenure in the given dataset? Provide a histogram and highlight the churned customers using a different color (Histogram is also called as column chart).

What is the most popular contract type among the customers in the given dataset? Provide a bar chart and highlight the churned customers using a different color.
What is the distribution of customers on each payment method in the given dataset?
Is there a relationship between total charges and customer tenure in the given dataset? Provide a scatter plot and highlight the churned customers using a different color.

What is the distribution of total revenue among the customers in the given dataset?